test(gateway): keep publication helpers in routed suite
